Removed all Sauce data, moved Item model into separate file
