+++ server/routes/index.js
@@ -1,9 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const itemRouter = require('./items')
 
 // different model routers
-router.use("/sauces", require("./sauces"));
 router.use("/items", require("./items"));
 router.use("/users", require("./users"));
 router.use("/orders", require("./orders"));
